pyproject.toml
[project]
name = "QuickApp-z7"
version = "7.2"
authors = [{name = "Andrea Censi", email = "AndreaCensi@users.noreply.github.com"}]
dependencies = [
    "ConfTools-z7",
    "compmake-z7",
    "numpy",
    "reprep-z7",
    "zuper-commons-z7",
    "zuper-params-z7",
    "zuper-utils-asyncio-z7",
    "zuper-zapp-z7",
    ]

[tool]
[tool.poetry]
name = "QuickApp-z7"
version = "7.2"
description = ""
authors = ["Andrea Censi <AndreaCensi@users.noreply.github.com>"]
packages = [
    {from = "src", include = "quickapp"},
    {from = "src", include = "quickapp/app_utils"},
    {from = "src", include = "quickapp/rm"},
    {from = "src", include = "reprep_quickapp"},
    ]

[tool.setuptools]
packages = [
    "quickapp",
    "quickapp.app_utils",
    "quickapp.rm",
    "reprep_quickapp",
    ]
package-dir = {"" = "src"}
